requirements.txt
bcrypt==4.0.1
blinker==1.6.2
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
email-validator==2.0.0.post2
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Email==1.4.4
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
Pillow==9.5.0
SQLAlchemy==2.0.13
svgwrite==1.4.3
Tree==0.2.4
typing_extensions==4.5.0
Werkzeug==2.3.4
WTForms==3.0.1
zipp==3.15.0
